+++ wasin/src/main/java/com/wasin/wasin/domain/dto/RouterResponse.java
@@ -1,5 +1,6 @@
 package com.wasin.wasin.domain.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class RouterResponse {
@@ -27,7 +28,7 @@ public class RouterResponse {
 
     public record Frame(Data data) {}
 
-    public record Data(List<List<Long>> values) {}
+    public record Data(List<List<BigDecimal>> values) {}
 
     public record Queries(List<Query> queries, String from, String to) {}
